Split off RSAPublicKey
